d_a_npc_cb1: Match leftover eventProc (demo partial)

+++ include/d/actor/d_a_player_npc.h
@@ -18,7 +18,7 @@ public:
     u32 isReturnLink() { return cLib_checkBit(m4E4, 0x01UL); }
     void onEventAccept() { cLib_onBit(m4E4, 0x02UL); }
     void offEventAccept() { cLib_offBit(m4E4, 0x02UL); }
-    u32 isEventAccept() { return cLib_checkBit(m4E4, 0x02UL); }
+    u8 isEventAccept() { return cLib_checkBit(m4E4, 0x02UL); }
     
     u8 getDamageFogTimer() { return mDamageFogTimer; }
     void setDamageFogTimer(u8 timer) { mDamageFogTimer = timer; }
